fix logger
